split the system/buffer into their own files, rename enum type
